backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
click==8.0.3
cloud-init==25.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==46.0.3
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.124.0
frozenlist==1.8.0
gpg==1.16.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.3
more-itertools==8.10.0
multidict==6.7.0
netifaces==0.11.0
oauthlib==3.2.0
passlib==1.7.4
pexpect==4.8.0
propcache==0.4.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1-modules==0.2.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-http-client==3.3.7
python-jose==3.5.0
python-magic==0.4.24
python-multipart==0.0.20
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.9.1
SecretStorage==3.3.1
sendgrid==6.12.5
service-identity==18.1.0
six==1.16.0
sos==4.5.6
SQLAlchemy==2.0.44
sqlmodel==0.0.27
ssh-import-id==5.11
starlette==0.50.0
systemd-python==234
twilio==9.8.8
Twisted==22.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.38.0
wadllib==1.3.6
Werkzeug==3.1.4
xkit==0.0.0
yarl==1.22.0
zipp==1.0.0
zope.interface==5.4.0
